Refactor propeller settings to use IntegerField for MCR operating point values, update data retrieval methods in UI components to ensure consistent access to system and user preferences, and adjust data refresh intervals across various dashboard views.

+++ src/db/models/propeller_setting.py
@@ -1,13 +1,13 @@
-from peewee import CharField, FloatField, BooleanField, Check
+from peewee import CharField, FloatField, BooleanField, Check, IntegerField
 from ..base import BaseModel
 
 
 class PropellerSetting(BaseModel):
     # MCR operation point
-    rpm_of_mcr_operating_point = FloatField(
+    rpm_of_mcr_operating_point = IntegerField(
         constraints=[Check('rpm_of_mcr_operating_point >= 0')])
 
-    shaft_power_of_mcr_operating_point = FloatField(
+    shaft_power_of_mcr_operating_point = IntegerField(
         constraints=[Check('shaft_power_of_mcr_operating_point >= 0')])
 
     # Normal propeller curve
